forgot to add some files

+++ codes/referee_questions/mcmc_new/mcmc_mpi/src/errors.c
@@ -11,11 +11,11 @@
 /* ----------------------------------------------------------------------- */
 
 /* Calculate chi2 for a process's given slice of pointings */
-double calculate_chi2(POINTING *p, int N_bins, int lower_ind, int upper_ind){
+long double calculate_chi2(POINTING *p, int N_bins, int lower_ind, int upper_ind){
 
     int i, j;
-    double chi2 = 0.0;
-    // double chi2_temp;
+    long double chi2 = 0.0;
+    // long double chi2_temp;
 
     for(i = lower_ind; i < upper_ind; i++){
 
@@ -36,8 +36,8 @@ double calculate_chi2(POINTING *p, int N_bins, int lower_ind, int upper_ind){
             //     / p[i].rbin[j].sigma2 );
             // chi2 += chi2_temp;
             // if (chi2_temp >1000){
-            //     fprintf(stderr, "chi2 is %lf for pointing %s, bin %d\n", chi2_temp, p[i].ID, j);
-            //     fprintf(stderr, "MM value is %lf\n", p[i].rbin[j].MM);
+            //     fprintf(stderr, "chi2 is %Lf for pointing %s, bin %d\n", chi2_temp, p[i].ID, j);
+            //     fprintf(stderr, "MM value is %Lf\n", p[i].rbin[j].MM);
             // }
         }
     }
